+++ contract/test/Nim.t.sol
@@ -87,7 +87,7 @@ contract NimTest is Test {
         game = nim.joinGame(gameId);
 
         // Store initial values
-        uint256 initialStones = game.rows[0];
+        uint8 initialStones = game.rows[0];
         
         // Player1 takes some stones
         vm.prank(player1);
@@ -129,8 +129,8 @@ contract NimTest is Test {
 
         // Test
         vm.prank(player1);
-        vm.expectRevert(abi.encodeWithSelector(Nim.InvalidRow.selector, gameId, uint256(game.rows.length)));
-        nim.playTurn(gameId, uint256(game.rows.length), 1);
+        vm.expectRevert(abi.encodeWithSelector(Nim.InvalidRow.selector, gameId, uint8(game.rows.length)));
+        nim.playTurn(gameId, uint8(game.rows.length), 1);
     }
 
     function test_playTurn_revert_invalidStones() public {
@@ -148,9 +148,9 @@ contract NimTest is Test {
             Nim.InvalidStones.selector,
             gameId,
             0,
-            uint256(game.rows[0] + 1)
+            uint8(game.rows[0] + 1)
         ));
-        nim.playTurn(gameId, 0, uint256(game.rows[0] + 1));
+        nim.playTurn(gameId, 0, uint8(game.rows[0] + 1));
     }
 
     function test_gameEndsWhenLastStoneTaken() public {
@@ -167,7 +167,7 @@ contract NimTest is Test {
         game = nim.playTurn(gameId, 0, game.rows[0]);
 
         // Take all stones from remaining rows
-        for (uint256 i = 1; i < game.rows.length; i++) {
+        for (uint8 i = 1; i < game.rows.length; i++) {
             if (i % 2 == 0) {
                 vm.prank(player1);
             } else {
